Create module starfit.storage.py
